Upgrade test runner to v2 (#57)
* Drop existing tests

* Add error tests for unitt

* Add failure tests for unitt

* Add success and misc tests for unitt

* Add parser tests

* Add parser logic

* Update Docker image and script to run

* Fix `sed` call

* Normalize exit codes for name errors on macOS

* Adjust error header normalization

* Use perl to unskip tests


* Stop script execution if not in temporary directory


* Add newline after parser tests


* Add missing return type annotation for parser/parser.py


* Simplify some multiline docstrings.

* Two newlines between functions and other expressions

* Import modules, not functions

* Single quotes to double quotes

* add macOS to parser GHA

* Create results.json if there are issues generating one normally.

* Apply code review

* Remove extra newline

* Update failure messages and file read/write handling

* Truncate message output to 65535 characters

* Only top-level mesages need to be capped

* Move `main`

* two spaces before comment

* Check for global fail status, not error status

* Move Unitt test result file tweaks into Python parser

* Revert "Move Unitt test result file tweaks into Python parser"

This reverts commit 275e428d920be8a2c8094fc481019b5a9225663d.

* Handle errors when creating directories, changing directories, or copying files

* Update parser/parser.py


* Update parser/parser.py


---------

+++ bin/run.sh
@@ -25,49 +25,45 @@ slug="$1"
 solution_dir=$(realpath "${2%/}")
 output_dir=$(realpath "${3%/}")
 results_file="${output_dir}/results.json"
+bin_dir=$(realpath "${0%/*}")
+parser_dir="${bin_dir}/../parser"
 
-# Create the output directory if it doesn't exist
 mkdir -p "${output_dir}"
 
 echo "${slug}: testing..."
 
-tmp_dir=$(mktemp -d -t "exercism-verify-${slug}-XXXXX")
-
-trap 'rm -rf "$tmp_dir"' EXIT
-cp -r "${solution_dir}/." "${tmp_dir}"
-cd "${tmp_dir}"
-test_file="tests/test-${slug}.art"
-sed -i -E 's/(test|it).skip/\1/g' "${test_file}"
-# Capture the output of the test run in case of failure or error
-test_output=$(arturo tester.art 2>&1)
-
-# Tweak the generated test file structure to valid Arturo code 
-sed -i \
-    -e 's/specs: \[/specs: @[/g' \
-    -e 's/tests: \[/tests: @[/g' \
-    -e 's/assertions: \[/assertions: @[@/g' ".unitt/tests/test-${slug}.art"
-
-# Check whether an error or failure occurred so we can pipe the captured output appropriately.
-cat > check-test-run-success.art << 'EOF'
-if empty? arg -> panic.unstyled.code:2 ""
-resultFile: ~".unitt/tests/test-|arg\0|.art"
-if not? file? resultFile -> panic.code:2 ""
-do resultFile
-if empty? specs -> panic.unstyled.code:1 ""
-testStatuses: flatten map specs 'describe ->
-                fold.seed:@[] describe\tests [acc test] ->
-                    append acc test\assertions\0\1
-if not? all? testStatuses -> panic.unstyled ""
-EOF
-
-arturo check-test-run-success.art "${slug}"
-result=$?
-if [ $result -eq 1 ]; then
-    jq -n --arg output "${test_output}" '{version: 1, status: "fail", message: $output}' > "${results_file}"
-elif [ $result -eq 2 ]; then
-    jq -n --arg output "${test_output}" '{version: 1, status: "error", message: $output}' > "${results_file}"
-else
-    jq -n '{version: 1, status: "pass"}' > "${results_file}"
+if ! tmp_dir=$(mktemp -d -t "exercism-verify-${slug}-XXXXX"); then
+    jq -n '{version: 2, status: "error", message: "The test runner failed to create the temporary directory for your test run. Please open a thread on the Exercism forums.", tests: []}' > "${output_dir}/results.json"
+    exit 1
 fi
 
-echo "${slug}: done"
+trap 'rm -rf "$tmp_dir"' EXIT
+if ! cp -r "${solution_dir}/." "${tmp_dir}"; then
+    jq -n '{version: 2, status: "error", message: "The test runner failed to copy your solution files. Please open a thread on the Exercism forums.", tests: []}' > "${output_dir}/results.json"
+    exit 1
+fi
+
+if ! cp "${parser_dir}/"*.py "${tmp_dir}/"; then
+    jq -n '{version: 2, status: "error", message: "The test runner failed to copy its internal parser files. Please open a thread on the Exercism forums.", tests: []}' > "${output_dir}/results.json"
+    exit 1
+fi
+
+if ! cd "${tmp_dir}"; then
+    jq -n '{version: 2, status: "error", message: "The test runner failed to change to the temporary directory. Please open a thread on the Exercism forums.", tests: []}' > "${output_dir}/results.json"
+    exit 1
+fi
+
+test_file="tests/test-${slug}.art"
+perl -i -pe 's/(test|it)\K.skip//g' "${test_file}"
+
+test_output=$(arturo tester.art 2>&1)
+test_filename=$(basename "${test_file}")
+result_art_file=".unitt/tests/${test_filename}"
+
+python3 parser.py "${test_file}" "${result_art_file}" "${test_output}"
+
+if [ ! -f results.json ]; then
+    jq -n '{version: 2, status: "error", message: "Critical error: The test runner failed to generate your results. Please open a thread on the Exercism forums and let us know.", tests: []}' > results.json
+fi
+
+cp results.json "${output_dir}/"
